watina: Querying and Processing Data from the INBO Watina Database

The R-package watina contains functions to query and process data from the Watina database at the Research Institute for Nature and Forest (INBO). This database primarily provides groundwater level and chemical data, mainly from natural areas in Flanders (Belgium).

Details

See the Index for an overview of the available functions. Basic data retrieval functions start with "get_".

Read the vignettes (linked by the Index) to get started with the package.

selectlocs_xg3 is a top-level function that allows to select locations based on user-specified conditions on XG3 availability and XG3 series.

selectlocs_chem is a top-level function that allows to select locations based on user-specified conditions on chemical data availability and properties.

For contributing, see the README on GitHub.

See also

Author

Maintainer: Floris Vanderhaeghe floris.vanderhaeghe@inbo.be (ORCID)

Other contributors:

Peter Desmet peter.desmet@inbo.be (ORCID) [contributor]

Jan Wouters jan.wouters@inbo.be (ORCID) [contributor]

Frederic Piesschaert frederic.piesschaert@inbo.be (ORCID) [contributor]

Mathias Wackenier mathias.wackenier@inbo.be (ORCID) [contributor]

Research Institute for Nature and Forest (INBO) info@inbo.be [copyright holder]