Convert a dataframe with X and Y coordinates to a geospatial points object

as_points is a convenience function which accepts as input a dataframe with X/Y coordinates (in meters), assumed to come from the coordinate reference system (CRS) 'Belge 1972 / Belgian Lambert 72' (EPSG 31370). It converts the dataframe into an sf points object in the same CRS.

as_points(df, xvar = "x", yvar = "y", remove = FALSE, warn_dupl = TRUE)

Arguments

df

A dataframe with X and Y coordinates in meters, assumed to be in the Belgian Lambert 72 CRS (EPSG-code 31370).

xvar

String. The X coordinate variable name. Defaults to "x".

yvar

String. The Y coordinate variable name. Defaults to "y".

remove

Logical. Should the X and Y coordinates be removed from the dataframe after conversion to a spatial object?

warn_dupl

Logical. Defaults to TRUE. Should the user be warned when duplicated coordinates are present in the result?

Value

An sf object of geometry type POINT with EPSG-code 31370 as coordinate reference system.

Details

As locations in Watina are typically defined by their X/Y coordinates, this function eases the conversion to spatial data. To later remove all spatial information from the result, you can use sf::st_drop_geometry().

Examples

library(tibble)
mydata <-
  tibble(
    a = runif(5),
    x = rnorm(5, 155763, 5),
    y = rnorm(5, 132693, 5)
  )
as_points(mydata)
#> Simple feature collection with 5 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 155762.1 ymin: 132690.2 xmax: 155767.7 ymax: 132692.5
#> Projected CRS: Belge 1972 / Belgian Lambert 72
#> # A tibble: 5 × 4
#>        a       x       y            geometry
#> *  <dbl>   <dbl>   <dbl>         <POINT [m]>
#> 1 0.0590 155762. 132692. (155762.1 132692.3)
#> 2 0.411  155766. 132692. (155765.5 132692.5)
#> 3 0.531  155765. 132692. (155764.9 132691.7)
#> 4 0.689  155768. 132690. (155767.7 132690.2)
#> 5 0.219  155764. 132693. (155763.8 132692.5)