Get locations from the database

Returns locations from the Watina database that meet several criteria (spatial or non-spatial), either as a lazy object or as a local tibble. Essential metadata are included in the result.

get_locs(con, max_filterdepth = 3, mask = NULL, join_mask = FALSE,
  buffer = 10, bbox = NULL, area_codes = NULL, loc_type = c("P",
  "S", "R", "N", "W", "D", "L", "B"), loc_validity = c("VLD", "ENT",
  "DEL", "CLD"), loc_vec = NULL, collect = FALSE)

Arguments

con
A DBIConnection object to Watina. See connect_watina to generate one.

max_filterdepth
Numeric. Maximum depth of the filter bottom below soil surface, as meters. This condition is only applied to piezometers.

mask
An optional geospatial filter of class sf. If provided, only locations that intersect with mask will be returned, with the value of buffer taken into account. The CRS must be Belgian Lambert 72 (EPSG-code 31370).

join_mask
Logical. Do you want to spatially join the attribute columns of mask to the resulting tibble? The spatial join is executed with st_intersects() as the topological operator. Beware: if the same location intersects with more than one element of mask (taking into account the value of buffer), that location will occur multiple times in the result. join_mask is ignored if mask is not provided.

buffer
Number of meters taken as a buffer to enlarge mask (or shrink it, if buffer < 0) if mask is provided.

bbox
Optional geospatial fiter (rectangle). A bounding box (class bbox), or a vector of four named elements xmin, xmax, ymin, ymax defining the boundary coordinates of a bounding box. If provided, only locations within this rectangular area will be returned. The CRS must be Belgian Lambert 72 (EPSG-code 31370).

area_codes
An optional vector with area codes. If provided, only locations within the areas will be returned.

loc_type
Type of the location (mainly: the type of measurement device). Defaults to "P", i.e. only groundwater piezometers are returned by default. Can be a vector with multiple selected values.

loc_validity
Validation status of the location. Defaults to "VLD", i.e. only validated locations are returned by default. Can be a vector with multiple selected values.

loc_vec
An optional vector with location codes. If provided, only locations are returned that are present in this vector.

collect
Should the data be retrieved as a local tibble? If FALSE (the default), a tbl_lazy object is returned (lazy query). Hence the result can be further built upon before retrieving data with collect().

Value

By default, a tbl_lazy object. With collect = TRUE or with a specified mask, a local tibble is returned.

(TO BE ADDED: Explanation on the variable names of the returned object)

Details

(TO BE ADDED: Explanation on the different available values of loc_type and loc_validity)

The lazy object returns a loc_wid variable, for further use in remote queries. However, don't use it in local objects: loc_wid is not to be regarded as stable. Therefore, collect = TRUE does not return loc_wid.

See also

Examples

if (FALSE) {
watina <- connect_watina()

get_locs(watina,
         bbox = c(xmin = 1.4e+5,
                  xmax = 1.7e+5,
                  ymin = 1.6e+5,
                  ymax = 1.9e+5))

get_locs(watina,
         area_codes = c("KAL", "KBR"),
         collect = TRUE)

get_locs(watina,
         area_codes = c("KAL", "KBR"),
         loc_type = c("P", "S"),
         collect = TRUE)

# Mark the different output of:
  get_locs(watina,
           loc_vec = c("KBRP081", "KBRP090", "KBRP095", "KBRS001"),
           loc_type = c("P", "S"),
           collect = TRUE)
  # versus:
  get_locs(watina,
           loc_vec = c("KBRP081", "KBRP090", "KBRP095", "KBRS001"),
           collect = TRUE)

# Selecting all piezometers with status VLD or ENT of the
# province "West-Vlaanderen":
data(BE_ADMIN_PROVINCE,
     package = "BelgiumMaps.StatBel")
library(dplyr)
library(sf)
library(stringr)
mymask <-
    st_as_sf(BE_ADMIN_PROVINCE) %>%
    filter(str_detect(TX_PROV_DESCR_NL, "West")) %>%
    st_transform(crs = 31370)
get_locs(watina,
         loc_validity = c("VLD", "ENT"),
         mask = mymask,
         buffer = 0)

# Disconnect:
DBI::dbDisconnect(watina)
}