watina: Querying & Processing Data From The INBO Watina Database (mainly groundwater data)

The R-package watina contains functions to query and process data from the Watina database at the Research Institute for Nature and Forest (INBO). This database provides groundwater level and chemical data, mainly from natural areas in Flanders (Belgium).

Details

See the Index for an overview of the available functions. Basic data retrieval functions start with "get_".

selectlocs_xg3 is a top-level function that allows to select locations based on user-specified conditions on XG3 availability and XG3 series.

For contributing, see the README on GitHub.

See also